feat: shift logging to __init__ again
